engine: Abstract out base functionalities for logging
