+++ invenio_logging/__init__.py
@@ -88,6 +88,6 @@ For more information about logging please see:
 
 from __future__ import absolute_import, print_function
 
-__version__ = "4.0.2"
+__version__ = "4.1.0"
 
 __all__ = ("__version__",)
