+++ setup.cfg
@@ -66,7 +66,6 @@ invenio_base.api_apps =
 invenio_search.index_templates =
     invenio_logging_datastream = invenio_logging.datastreams.templates
 invenio_logging.datastreams.builders =
-    audit = invenio_logging.datastreams.audit_logs:AuditLogBuilder
     app = invenio_logging.datastreams.app_logs:AppLogBuilder
 invenio_search.component_templates =
     logging = invenio_logging.datastreams.component_templates
