+++ tests/datastreams/test_manager.py
@@ -9,8 +9,8 @@
 
 import pytest
 
-from invenio_logging.datastreams.log_event import LogEvent
-from invenio_logging.datastreams.managers import LogManager
+from invenio_logging.engine.log_event import BaseLogEvent
+from invenio_logging.engine.managers import LogManager
 
 
 class MockLogBuilder:
@@ -41,7 +41,7 @@ def test_log_event():
     mock_builder = MockLogBuilder()
     log_manager.register_builder("info", mock_builder)
 
-    log_event = LogEvent(
+    log_event = BaseLogEvent(
         log_type="info",
         message="Test event",
         event="test",
@@ -53,5 +53,5 @@ def test_log_event():
 def test_log_event_no_builder():
     log_manager = LogManager()
     with pytest.raises(ValueError):
-        log_event = LogEvent(log_type="error", message="Test event", event="test")
+        log_event = BaseLogEvent(log_type="error", message="Test event", event="test")
         log_manager.log(log_event, async_mode=False)
